♻️파일분리
